Added a web app cookie that tracks a user session
This is a unsecure implementation of user sessions.

A user session is determined by a cookie set in thier browser on inital visit.

User's can only see, interact with and delete thier own uploaded images.

I added a delete images button to the main page since it now only affects the current user.

This is considered unsecure since the cookie is just a random string and users can access other
users images with thier cookie, similarly anyone with cluster access can see all user images.

To Do:

This commit should be followed up with the addition of scheduled cleanup of old user files